requirements.txt
anyio==4.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
python-dotenv==1.0.1
python-telegram-bot==21.4
requests==2.32.3
schedule==1.2.2
sniffio==1.3.1
typeguard==4.3.0
typing_extensions==4.12.2
urllib3==2.2.2
